fixed light issues

+++ gameObjects.js
@@ -638,8 +638,8 @@ function LightEmitting(lightlevel) {
 		this.light = light;
 	}
 	this.getLight = function() {
-	  if (parseInt(this.light) === "NaN") { alert(this.getName() + " has lightlevel that is NaN."); }
-		return parseInt(this.light);
+	  if (parseFloat(this.light) === "NaN") { alert(this.getName() + " has lightlevel that is NaN."); }
+		return parseFloat(this.light);
 	}
 }
